% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_posterior.R
\name{draw_posterior_quap}
\alias{draw_posterior_quap}
\title{Posterior samples using \code{rethinking::extract.samples} as
\code{draws_rvars}}
\usage{
draw_posterior_quap(obj, n = 1L)
}
\arguments{
\item{obj}{S4 object of class \code{map}.}

\item{n}{Sample size.}
}
\value{
\code{draws_rvars} object
}
\description{
Posterior samples using \code{rethinking::extract.samples} as
\code{draws_rvars}.
}
\details{
get posterior samples from a \code{map} object formatted as a \code{draws_rvars}
object to be used by the \code{posterior} package.
}
