% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflRethinking.R
\docType{package}
\name{eflRethinking}
\alias{eflRethinking}
\title{eflRethinking: A package with functions to use with Rethinking, 2nd edition}
\description{
The eflRethinking package provides functions mainly to solve practices
of Rethinking, 2nd edition
}
