% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_posterior.R
\name{gt_posterior}
\alias{gt_posterior}
\title{Format a gt table for posterior summary}
\usage{
gt_posterior(
  df,
  var_col = "var",
  labs = list(title = "Posterior Summary", subtitle = "Practice"),
  qtl = NULL,
  .data_color = "mintcream",
  .tab_options = NULL,
  digits = 2L
)
}
\arguments{
\item{df}{Dataframe. Posterior summary.}

\item{var_col}{Name of the column with the variables.}

\item{labs}{List with the following items}

\item{qtl}{List used to change the name of the quantile columns. Should
be formatted as, for example.}

\item{.data_color}{Data color used by \code{gt::data_color()}}

\item{.tab_options}{List of tab_options with the following defaults
\describe{
  \item{data}{"mintcream"}
  \item{heading.background.color}{"lavender"}
  \item{heading.title.font.weight}{"bold"}
  \item{heading.subtitle.font.weight}{"bold"}
  \item{column_labels.background.color}{"oldlace"}
  \item{column_labels.font.weight}{"bold"}
  \item{stub.background.color}{"oldlace"}
  }}

\item{digits}{Integer. Digits used to format data numbers.}
}
\value{
gt table of posterior summary
}
\description{
Format a gt table for posterior summary. Usually from \code{quap},
\code{brm} and \code{inla}.
}
\details{
The format will rename the quantiles based on the \code{qtl} argument which
must have the same number of element as the corresponding columns.
}
