% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_posterior.R
\name{gt_posterior_compare}
\alias{gt_posterior_compare}
\title{Create a gt table comparing posteriors}
\usage{
gt_posterior_compare(
  dfs,
  var_df = names(dfs)[1],
  var_col = "var",
  labs = list(title = "Posterior Summary Comparison", subtitle = "Practice"),
  .data_color = "mintcream",
  .tab_options = NULL,
  digits = 2L
)
}
\arguments{
\item{dfs}{List of dataframes with posterior summaries.}

\item{var_df}{Name of the summary to use as a reference for variable names.}

\item{var_col}{Name of the column with the variables.}

\item{labs}{List with the following items}

\item{.data_color}{Data color used by \code{gt::data_color()}}

\item{.tab_options}{List of tab_options with the following defaults
\describe{
  \item{"data"}{"mintcream"}
  \item{"heading.background.color"}{"darkslategray"}
  \item{"heading.title.font.weight"}{"bold"}
  \item{"heading.subtitle.font.weight"}{"bold"}
  \item{"column_labels.background.color"}{"seashell"}
  \item{"column_labels.font.weight"}{"bold"}
  \item{"stub.background.color"}{"seashell"}
  }}

\item{digits}{Integer. Digits used to format data numbers.}
}
\value{
gt table of posterior summary comparison
}
\description{
Create a gt table comparing posteriors. Usually from \code{quap}, \code{brm}
and \code{inla}.
}
\details{
The list of summaries must meet these conditions
\itemize{
  \item The summary called "var_df" in the list will be used to name the rows
  \item All summaries must have the rows in the same order
  \item All summaries must have the same column names
  }
}
