% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflRethinking.R
\docType{package}
\name{eflRethinking}
\alias{eflRethinking}
\title{tUilities to use with the `Rethinking` package}
\description{
Utilities to use with the `Rethinking` package.
}
\details{
Utilities to use with the `Rethinking` package.
The functions are mainly used to convert objects so that they have
a `brms` feel, and, in particular, can be used as `rvars` and `draws_vars`
with the `posterior` package.
}
